Refactor QuickSight helper
